+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 6/17/2023
 - Major: Added Telegram users now can use commands to check stamina, and check expedition status by using `/stamina` and `/expedition` respectively.
 - Minor: Added `prepareMessage` function to prepare message for Telegram and Discord
+- Minor: Typo fix for `expidition` to `expedition` at expedition cron
 
 ## 6/16/2023
 - Minor: Discord embed generator now accept `stamina` and `expedition` as a parameter to generate the embed